### 2022-06-15 fbdb99a
docs: add GPL3 licence file
