### 2022-06-15 332efcd
docs: modify invalid DOIs and add missing
