+++ paper.bib
@@ -46,7 +46,7 @@ volume = {33},
 number = {5},
 pages = {1151-1163},
 keywords = {accessibility, geographic bias, maps, publication density, sampling effort, accesibilidad, densidad de publicaciones, esfuerzo de muestreo, mapas, sesgo geográfico, 地理偏倚, 地图, 发表密度, 抽样工作},
-doi = {https://doi.org/10.1111/cobi.13326},
+doi = {10.1111/cobi.13326},
 url = {https://conbio.onlinelibrary.wiley.com/doi/abs/10.1111/cobi.13326},
 eprint = {https://conbio.onlinelibrary.wiley.com/doi/pdf/10.1111/cobi.13326},
 year = {2019}
@@ -93,7 +93,7 @@ volume = {118},
 pages = {102526},
 year = {2022},
 issn = {1569-190X},
-doi = {https://doi.org/10.1016/j.simpat.2022.102526},
+doi = {10.1016/j.simpat.2022.102526},
 url = {https://www.sciencedirect.com/science/article/pii/S1569190X22000296},
 author = {Matthew Snyder and Byran J. Smucker},
 keywords = {Emergency medical services, Discrete event simulation, Metamodeling, Particle swarm optimization, Random forest},
@@ -165,7 +165,7 @@ volume = {191},
 pages = {103452},
 year = {2019},
 issn = {0169-2046},
-doi = {https://doi.org/10.1016/j.landurbplan.2018.10.008},
+doi = {10.1016/j.landurbplan.2018.10.008},
 url = {https://www.sciencedirect.com/science/article/pii/S0169204618311083},
 author = {JieLan Xu},
 }
@@ -206,7 +206,7 @@ volume = {97},
 pages = {102964},
 year = {2021},
 issn = {1361-9209},
-doi = {https://doi.org/10.1016/j.trd.2021.102964},
+doi = {10.1016/j.trd.2021.102964},
 url = {https://www.sciencedirect.com/science/article/pii/S1361920921002625},
 author = {Maxime Savaria and Philippe Apparicio and Mathieu Carrier},
 keywords = {Filtered permeability, Transport planning, Bicycle, Sustainable urban mobility},
@@ -244,7 +244,8 @@ DOI = {10.3390/ijgi8090400}
   journal={International Journal of Data Science and Analytics},
   pages={1--16},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s41060-022-00328-x}
 }
 
 @article{CHEN2021126,
@@ -254,7 +255,7 @@ volume = {222},
 pages = {126-136},
 year = {2021},
 issn = {0002-9394},
-doi = {https://doi.org/10.1016/j.ajo.2020.08.038},
+doi = {10.1016/j.ajo.2020.08.038},
 url = {https://www.sciencedirect.com/science/article/pii/S0002939420304670},
 author = {Evan M. Chen and Aneesha Ahluwalia and Ravi Parikh and Kristen Nwanyanwu},
 }
@@ -283,7 +284,7 @@ volume = {61},
 pages = {127097},
 year = {2021},
 issn = {1618-8667},
-doi = {https://doi.org/10.1016/j.ufug.2021.127097},
+doi = {10.1016/j.ufug.2021.127097},
 url = {https://www.sciencedirect.com/science/article/pii/S1618866721001229},
 author = {Elizabeth Walker and Dimitrios Bormpoudakis and Joseph Tzanopoulos},
 keywords = {Children, Distance analysis, Green spaces, Nature, Planning strategies},
@@ -294,14 +295,16 @@ keywords = {Children, Distance analysis, Green spaces, Nature, Planning strategi
 
 
 @Article{stplanr,
[…]
