### 2022-06-15 #4477 @rCarto
Hello,
I've added a license file and corrected missing or invalid DOIs.
I suggest wcjochem, mikemahoney218, josiahparry, ssujit or Maczokni  as potential reviewers.
