### 2022-06-17 #4477 @elbeejay
Hi @rCarto, thanks for making this submission. I'll have a closer look this weekend before making any pre-review suggestions and searching for reviewers, but when I took a glance I had one question. In the README, you state:

> :heavy_exclamation_mark: **To consider when using OSRM**:
\> [“Most of the previously active core devs have either moved on to new
roles, or are simply busy on different projects (…)”](https://github.com/Project-OSRM/osrm-backend/issues/5463)

Just to be clear, you are planning to maintain `osrm` (the submitted package), but the upstream dependency OSRM is no longer being actively maintained? It is a bit confusing given the similar names. Will the lack of upstream maintenance pose a problem in the near-future, and do you know of alternative routing services you might be able to use instead (or any under development)? Or is the package written in a way that it can flexibly point to a variety of routing services?

Thanks and let me know if you have any questions for me!

Jay
