### 2022-06-18 #4477 @rCarto
Hi @elbeejay ,

> Just to be clear, you are planning to maintain `osrm` (the submitted package), but the upstream dependency OSRM is no longer being actively maintained?

I plan to maintain `osrm` the R package. I've placed this warning in the README at a time I thought there could be some problems with OSRM maintenance. It has appeared since that the project is still maintained, it receives contributions, its issues are taken care of, its mailing list is still active and new versions are released. In fact I should just remove this statement from the README.

> Will the lack of upstream maintenance pose a problem in the near-future, and do you know of alternative routing services you might be able to use instead (or any under development)? Or is the package written in a way that it can flexibly point to a variety of routing services?

The package is written for this software API only.
