### 2022-06-19 #4477 @elbeejay
Thanks @rCarto that explanation helps.

Some minor suggestions for the paper:

- In the second sentence of the Summary section add "users" (see below)
> This package allows **users** to compute routes, trips, isochrones and travel distances matrices (travel times and kilometric distances) based on the OSM road network.

- Consider changing the final sentence of the Summary section to one of the below versions (or similar):
> An instance of OSRM can be installed on a local or remote server, allowing free and heavy usage of the routing engine.

> An instance of OSRM can be installed on a local or remote server and allows free and heavy usage of the routing engine.

- The second sentence in the Statement of need is very casual, I'd consider finding a way to write it more succinctly, maybe something like: "The simplest distance metric is the Euclidean distance (or distance as the crow flies), it is easy and inexpensive to compute."

- The "a" before the word "research" on line 23 should be deleted. Consider revising that sentence to something like the following:
> But whenever research aims at studying human activities (transport of persons or goods for example) it is common to use more realistic metrics based on road distance or travel time.

- In the sentence referencing uses of the package in various fields, I would suggest switching "environmental science" for "environment" and ending the sentence with "and linguistics."

- In line 38 the "a" before "suggested" should be deleted

- In line 45 I don't know if you mean to use "implies" I think something like "imposes" might be more appropriate.

- For the State of the field section it looks like the line breaks in the `paper.md` file are being carried over into the PDF for some reason. Can you try putting all of the text on a continuous line without breaks?

Please feel free to accept, reject, or question any of my above suggestions. I hope they all make sense - I would prefer to see at least some minor revisions made to the paper before we have reviewers spending time looking at it.

Thanks,
Jay
