### 2022-06-19 9097701
docs: remove comment on OSRM maintenance, rebuild README
