### 2022-06-19 #4477 @rCarto
Thank you very much @elbeejay !
I have accepted all your suggestions. As you may have noticed, I am not a native English speaker. I have also removed the warning about OSRM maintenance in the README.
