### 2022-06-20 #4477 @elbeejay
@rCarto thank you for making those changes. Looking at the package again, I am going to ask the broader editorial team if this fits within the current scope for JOSS papers. Since your last JOSS publication our guidelines have changed a bit (see documentation on [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) and the blog post ["Minimum publishable unit"](https://blog.joss.theoj.org/2020/07/minimum-publishable-unit)).

I personally am on the fence as this is an API wrapper for OSRM, so the package is a bit small as it just provides 5 functions around that API. However on the positive side, you cited a number of different publications from a variety of fields that have made use of `osrm`, so it is clearly being used by the broader scientific community. In an effort to be cautious, I just want to get the feedback from the rest of the editorial team before continuing.

Thank you for your patience and understanding, I'm hoping to hear back from the other editors in a about a week or so. I will provide an update then.
