### 2022-06-20 #4477 @rCarto
@elbeejay ,
I understand your concern.
Though, I would like to add two notes:
- I'm not sure that the package is that "thin" around the OSRM API,
- 2 of its functions (`osrmIsochrone()` and `osrmIsometric()`) provide original features not available through OSRM alone.

Anyway, I would understand if the package is considered outside the scope of JOSS.
