### 2022-06-20 #4477 @elbeejay
Thank you for the clarification that the two functions provide original functionality, I will add that to my note to the other editors.
