### 2022-07-07 a3b3bc0
docs: update NEWS & README
