### 2022-07-17 #4574 @editorialbot
[REVIEW]: osrm: Interface Between R and the OpenStreetMap-Based Routing Service OSRM
**Submitting author:** @rCarto (<a href="http://orcid.org/0000-0002-1932-3323">Timothée Giraud</a>)
**Repository:** https://github.com/riatelab/osrm/
**Branch with paper.md** (empty if default branch):
**Version:** v4.0.0
