### 2022-07-17 #4574 @elbeejay
@mikemahoney218  and @JosiahParry thanks again for agreeing to review this submission to JOSS. We've just switched over from whedon to the new editorialbot, and the review process has changed subtly. Please refer to the post at the top of this issue page for a link to the reviewer guidelines as well as the command for generating your reviewer checklist (@mikemahoney218 wasted no time in generating this :confetti_ball:!) -- please feel free to reach out to me with any questions you may have. Please open issues related to this JOSS review in the [osrm repository](https://github.com/riatelab/osrm/) itself, and "link" them to this review issue by posting this URL (https://github.com/openjournals/joss-reviews/issues/4574) in the text of the issue you open.

Thanks again for reviewing for JOSS, I'm going to tell the editorial bot to send out review reminders in 3 weeks just so we don't collectively let this slip through the cracks.
