### 2022-07-18 c3720a2
fix: more efficient use of st_coordinates()
fix #91
