### 2022-08-05 #4574 @elbeejay
Hello @wcjochem @JosiahParry @mikemahoney218 ! I just wanted to check-in and make sure the reviews are going okay and that this hasn't fallen off of anyone's radar. Let me know if you have any questions about the process. As a reminder please make review comments as __issues__ in the `osrm` repository with a link to this review issue so everything stays connected.

Thanks!
Jay
