### 2022-08-23 riatelab/osrm#92 @mikemahoney218
Software paper
# Software Paper

_This issue is posted as part of my review of the software paper for JOSS, at https://github.com/openjournals/joss-reviews/issues/4574_

Good job on the paper! I have one comment under "State of the Field" which I think needs to be resolved, and then a few suggestions for improvements which are entirely up to you (but wouldn't stop me from recommending acceptance).

## State of the Field

This is the only box I haven't "checked" on the review checklist for this section: I think it'd be useful in your "State of the Field" section to discuss the `osrmr` package, which also provides hooks to the OSRM API:
https://cran.r-project.org/web/packages/osrmr/index.html

I think a discussion of the differences in implementations would be sufficient for me to check off this box.

## Other Suggestions

I wanted to mention a few potential improvements I think you could make, which I think will make the paper more useful for others going forward, but I don't view these as necessary before the paper can be accepted.

First and foremost, I think the paper would benefit from a short section on how to use the package. This isn't a requirement of JOSS, but seems to be pretty standard across at least the R papers I've read in the journal (see for instance the first few entries at https://joss.theoj.org/papers/in/R ). Adding a short demonstration of key functionality would be useful.

I also have a handful of phrasing suggestions for the paper:

https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/paper.md?plain=1#L23
I'd personally write "It contains five functions that interface _with_ OSRM services:"

https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/paper.md?plain=1#L37
Consider: "...which is easy and inexpensive to compute" (instead of "it is")

https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/paper.md?plain=1#L59
I'd personally change "exploitation" to "analysis".
