### 2022-08-23 riatelab/osrm#93 @mikemahoney218
Functionality & Documentation
# Functionality & Documentation
_This issue is posted as part of my review of the software paper for JOSS, at https://github.com/openjournals/joss-reviews/issues/4574_

There are two related items on the JOSS reviewer checklist I don't think I'm able to check off yet. First, under Functionality:

> Functionality: Have the functional claims of the software been confirmed?

And second, under Documentation:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Running through the test suite, it seems like the majority of tests are testing for (1) queries are executed successfully and (2) the return objects are sf objects. I didn't see unit testing of the utility functions not directly querying the OSRM API, or for correctness of the returned object.

I understand not wanting to test against results from the OSRM API, because you aren't responsible for changes in their data or routing methods. However, I'm wondering if it's possible to directly test the functions in `utils.R` to ensure their outputs don't accidentally change as the result of a different fix. I think that would make it easier to assess what functionality is "guaranteed" by this package. I think adding these tests would satisfy both items on the checklist.

A bigger suggestion, which I don't think is _necessary_ for this submission but I do think would be helpful, would be to extract some processing steps into their own functions and test those. For instance, the following section could be extracted into its own function and then tested directly (using data from a request saved in `inst`), making it easier to verify that the computations done by the package itself are stable and correct:
https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/R/osrmIsochrone.R#L139-L155
