### 2022-08-24 riatelab/osrm#94 @JosiahParry
osrmTable does not accept sfc objects
In testing `osrm` I discovered that it cannot handle sfc object as inputs. I'd suggest that sfc object be a possible input for both `src` and `dst`.

Additionally, I'd recommend including object type checking as the error is uninformative. There is no `lon` argument so the user is left to wonder what they did wrong without appropriate guidance.

``` r
library(sf)
#> Linking to GEOS 3.9.1, GDAL 3.2.3, PROJ 7.2.1; sf_use_s2() is TRUE
library(osrm)
#> Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
#> Routing: OSRM - http://project-osrm.org/

geo <- st_centroid(sfdep::guerry) |>
  st_geometry()
#> Warning in st_centroid.sf(sfdep::guerry): st_centroid assumes attributes are
#> constant over geometries of x

# set appropriate crs
st_crs(geo) <- 27572

osrmRoute(src = geo[1],
          dst = geo[2],
          overview = "full")
#> Error in input_route(x = src, id = "src", single = TRUE): object 'lon' not found

osrmRoute(
  src = st_as_sf(geo[1]),
  dst = st_as_sf(geo[2]),
          overview = "full"
  )
#> Simple feature collection with 1 feature and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 688513.5 ymin: 2125991 xmax: 847686.6 ymax: 2513124
#> Projected CRS: NTF (Paris) / Lambert zone II
#>     src dst duration distance                       geometry
#> 1_1   1   1 326.6383 529.2747 LINESTRING (832834.2 212655...
```
