### 2022-08-24 riatelab/osrm#95 @JosiahParry
osrmTable does not accept sfc objects as input
Similar to #94. It's important to allow users to provide sfc class objects that contain only geometry. sfc objects are under the hood lists where elements are a geometry. You can use the element indix similar to how you use the row name in data frames as the id.

If you elect not to, I'd suggest having type checking and informative error handling. The provided error is generated inside the body of the function and is not created by any error handling

``` r
library(sf)
#> Linking to GEOS 3.9.1, GDAL 3.2.3, PROJ 7.2.1; sf_use_s2() is TRUE
library(osrm)
#> Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
#> Routing: OSRM - http://project-osrm.org/

geo <- st_centroid(sfdep::guerry) |>
  st_geometry()
#> Warning in st_centroid.sf(sfdep::guerry): st_centroid assumes attributes are
#> constant over geometries of x

# set appropriate crs
st_crs(geo) <- 27572

osrmTable(src = geo[1],
          dst = geo[2])
#> Error in names(src) <- c("id", "lon", "lat"): 'names' attribute [3] must be the same length as the vector [1]

osrmTable(
  src = st_as_sf(geo[1]),
  dst = st_as_sf(geo[2])
)
#> $durations
#>       1
#> 1 326.6
#>
#> $sources
#>        lon      lat
#> 1 5.348519 46.09909
#>
#> $destinations
#>        lon      lat
#> 1 3.558656 49.55887
```
