### 2022-08-24 riatelab/osrm#96 @JosiahParry
incorrect documentation: osrm does not require WGS84
One of the undocumented features of `osrm` is that projected coordinate systems are first transformed to wgs84.
The documentation states that geometry _must_ be lon lat. This is false.

https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/R/osrmTable.R#L6

There are a few possible ways to ameliorate this that I can see.

1. emit a message or warning upon transformation
2. add a details section to the osrm functions which document this practice
3. do both :)

The latter would be my suggestion
