### 2022-08-24 riatelab/osrm#97 @JosiahParry
osrmRoute does not require wgs84
based on #96.

 `osrmRoute()` when provided non-wgs84 CRS returns the provided CRS.  It is transformed upon ingest and output. For osrmRoute I'd likely either remove the wgs84 requirement or document the transformation process that occurs.

``` r
library(sf)
#> Linking to GEOS 3.9.1, GDAL 3.2.3, PROJ 7.2.1; sf_use_s2() is TRUE
library(osrm)
#> Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
#> Routing: OSRM - http://project-osrm.org/

geo <- st_centroid(sfdep::guerry) |>
  st_geometry()
#> Warning in st_centroid.sf(sfdep::guerry): st_centroid assumes attributes are
#> constant over geometries of x

# set appropriate crs
st_crs(geo) <- 27572

# transform to wgs84
geo2 <- st_transform(geo, 4326) |>
  st_as_sf()

r1 <- osrmRoute(
  src = st_as_sf(geo[1]),
  dst = st_as_sf(geo[2])
)

r2 <- osrmRoute(src = geo2[1,],
                dst = geo2[2,])

identical(r1, r2)
#> [1] FALSE
```
