### 2022-08-24 riatelab/osrm#98 @JosiahParry
osrmTrip does not accept sfc class objects
similar to #94 and #95

osrmTrip does not accept sfc class objects.
The error is generated by the OSRM api and not by the function itself. It would be great if the function could accept an sfc class object. Additionally, the function should type check objects.

``` r
library(sf)
#> Linking to GEOS 3.9.1, GDAL 3.2.3, PROJ 7.2.1; sf_use_s2() is TRUE
library(osrm)
#> Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
#> Routing: OSRM - http://project-osrm.org/

geo <- st_centroid(sfdep::guerry) |>
  st_geometry()
#> Warning in st_centroid.sf(sfdep::guerry): st_centroid assumes attributes are
#> constant over geometries of x

# set appropriate crs
st_crs(geo) <- 27572

# transform to wgs84
geo2 <- st_transform(geo, 4326) |>
  st_as_sf()

osrmTrip(geo)
#> Error: OSRM API request failed [400]
#> InvalidQuery
#> Query string malformed close to position 17

osrmTrip(geo2[1:5,])
#> [[1]]
#> [[1]]$trip
#> Simple feature collection with 5 features and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 2.431282 ymin: 44.04703 xmax: 6.26198 ymax: 49.60789
#> Geodetic CRS:  WGS 84
#>   start end duration distance                       geometry
#> 1     1   2 326.6383 529.2747 LINESTRING (5.348519 46.099...
#> 2     2   3 331.0183 470.3772 LINESTRING (3.558656 49.558...
#> 3     3   4 400.5667 463.9863 LINESTRING (3.187705 46.394...
#> 4     4   5 153.9617 125.8599 LINESTRING (6.24295 44.1053...
#> 5     5   1 251.1283 265.9457 LINESTRING (6.26198 44.6634...
#>
#> [[1]]$summary
#> [[1]]$summary$duration
#> [1] 1463.313
#>
#> [[1]]$summary$distance
#> [1] 1855.444
```
