### 2022-08-24 riatelab/osrm#99 @JosiahParry
osrmTrip accepts polygon input
`osrmTrip()` accepts a polygon input and returns an output.  It is unclear as to what is happening under the hood for this to happen. The results are different than the output of `osrmTrip()` on using the centroid.

This function should likely error out when providing polygons. If not, there should be a warning emitted that informs the user of what is happening.

``` r
library(sf)
#> Linking to GEOS 3.9.1, GDAL 3.2.3, PROJ 7.2.1; sf_use_s2() is TRUE
library(osrm)
#> Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
#> Routing: OSRM - http://project-osrm.org/

geo <- st_centroid(sfdep::guerry) |>
  st_geometry()
#> Warning in st_centroid.sf(sfdep::guerry): st_centroid assumes attributes are
#> constant over geometries of x

# set appropriate crs
st_crs(geo) <- 27572

geo3 <- sfdep::guerry
st_crs(geo3) <- 27572

osrmTrip(st_sf(geo))
#> [[1]]
#> [[1]]$trip
#> Simple feature collection with 85 features and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 121674.4 ymin: 1733111 xmax: 991123.6 ymax: 2610991
#> Projected CRS: NTF (Paris) / Lambert zone II
#> First 10 features:
#>    start end  duration distance                       geometry
#> 1      1  69  94.87833 113.4403 LINESTRING (832834.1 212655...
#> 2     69  37 116.39667 108.8712 LINESTRING (768820.9 218511...
#> 3     37  19 104.44667 129.2257 LINESTRING (856787.7 219751...
#> 4     19  50 121.98667 137.8112 LINESTRING (783424.3 227225...
#> 5     50  84 101.12667 113.0582 LINESTRING (815192.8 234955...
#> 6     84  68  89.42000 101.1851 LINESTRING (900579.5 236301...
#> 7     68  23  85.25667  90.4215 LINESTRING (881484.6 230043...
#> 8     23  66 106.37000 135.5095 LINESTRING (905132.9 224839...
#> 9     66  65  88.27333 112.1390 LINESTRING (969137 2329307,...
#> 10    65  55  94.72667 114.8728 LINESTRING (983968.1 242075...
#>
#> [[1]]$summary
#> [[1]]$summary$duration
#> [1] 8435.987
#>
#> [[1]]$summary$distance
#> [1] 9690.122
osrmTrip(geo3)
#> [[1]]
#> [[1]]$trip
#> Simple feature collection with 85 features and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 121674.4 ymin: 1733111 xmax: 991123.6 ymax: 2610991
#> Projected CRS: NTF (Paris) / Lambert zone II
#> First 10 features:
#>    start end  duration distance                       geometry
#> 1      1  37  88.02500 106.0664 LINESTRING (832834.1 212655...
#> 2     37  19 100.98667 127.8031 LINESTRING (856788.6 219750...
#> 3     19  50 118.46167 136.1007 LINESTRING (784190.5 227225...
#> 4     50  84 101.12667 113.0582 LINESTRING (815192.8 234955...
#> 5     84  68  89.42000 101.1851 LINESTRING (900579.5 236301...
#> 6     68  23  85.25667  90.4215 LINESTRING (881484.6 230043...
#> 7     23  66 106.37000 135.5095 LINESTRING (905132.9 224839...
#> 8     66  65  88.27333 112.1390 LINESTRING (969137 2329307,...
#> 9     65  55  94.72667 114.8728 LINESTRING (983968.1 242075...
#> 10    55  52  70.68000  71.1266 LINESTRING (915870.5 245764...
#>
#> [[1]]$summary
#> [[1]]$summary$duration
#> [1] 8411.185
#>
#> [[1]]$summary$distance
#> [1] 9577.701
```
