### 2022-08-24 riatelab/osrm#100 @JosiahParry
osrmTrip summary list element unclear
https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/R/osrmTrip.R#L23

It is unclear what the summary list element is. While I know they are durations and distances, I am unclear if it is the total duration and distance for all returned routes.
