### 2022-08-24 riatelab/osrm#101 @JosiahParry
Typo line 22
https://github.com/riatelab/osrm/blob/c3720a27dc23c162ff93e224e0ffc50abfb7fbb0/paper.md?plain=1#L22

"This package allows to compute routes" is grammatically incorrect. A possible alternative would be "this package enables the computation of routes..."
