### 2022-08-24 riatelab/osrm#102 @JosiahParry
Substantial Scholarly Effort
In reviewing `osrm` for JOSS, I am unsure if it meets the requirements for substantial scholarly effort. [JOSS requirements](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) make note of what constitutes substantial scholarly effort. Some of these include:

- "Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected."
- "requires that software should be feature-complete "
- "'thin' API clients, and single-function packages are not acceptable."

### Total lines of code

Regarding the first point, the entire package contains 1020 lines of code which includes package documentation in the form of roxygen comments. When excluding them, the package comes out to a total of 654 lines of code.

``` r
# taken from https://stackoverflow.com/questions/47330367/get-the-number-of-lines-of-code-in-a-r-script
number_of_lines_of_code <- function(file_path){
  file <- readr::read_file(file_path)
  file_lines <- file |> stringr::str_split("\n")
  first_character_of_lines <- file_lines |>
    lapply(function(line)stringr::str_replace_all(line," ","")) |>
    lapply(function(line)stringr::str_sub(line,1,1)) |>
    unlist()
  sum(first_character_of_lines != "#" & first_character_of_lines != "\r")
}

burl <- "https://raw.githubusercontent.com/riatelab/osrm/master/R"

file.path(burl, c(
  "osrmIsochrone.R",
  "osrmRoute.R",
  "osrmTable.R",
  "osrmTrip.R",
  "package.R",
  "utils.R",
  "zzz.R"
)) |>
  purrr::map_dbl(number_of_lines_of_code) |>
  sum()
#> [1] 654
```

### Software is feature complete

The package contains 5 functions. I've encountered a few problems that I think should be rectified before it is considered feature complete (see issues #94 - #99). Noteably there is insufficient error handling and type checking and unnecessary tacit assumptions of input objects.

### Thin API clients

This package contains 5 functions 3 of which I think can be considered fairly thin interfraces to the osrm API. The 2 functions which I think can be considered substantial are `osrmIsochrone()` and `osrmIsometric()` as these do more than just type convert and send http requests.
