### 2022-08-24 riatelab/osrm#103 @JosiahParry
no contributing guidelines
Per JOSS requirements, there needs to be "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support." At present, I cannot find such guidelines. Consider using `usethis::use_tidy_contributing()` to include guidelines or write your own `CONTRIBUTING.md`
