### 2022-08-24 #4574 @JosiahParry
I have reviewed this package and at this moment am unable to check off the sections `Substantial Scholarly Effort`, `Functionality Documentation`, `Automated Test`,  `Community Guidelines`, and `Quality of the writing`.

re substantial scholarly effort: https://github.com/riatelab/osrm/issues/102
re functionality documentation: https://github.com/riatelab/osrm/issues/100, https://github.com/riatelab/osrm/issues/97, and https://github.com/riatelab/osrm/issues/96
re automated testing: I agree with @mikemahoney218's statement regarding testing suite in https://github.com/riatelab/osrm/issues/93.
re community guidelines: https://github.com/riatelab/osrm/issues/103
re quality of the writing: i also concur with @mikemahoney218 in https://github.com/riatelab/osrm/issues/92, additionally there is at least one typo that I cought https://github.com/riatelab/osrm/issues/101
