### 2022-08-24 #4574 @elbeejay
Thank you to both @mikemahoney218 and @JosiahParry for reviewing the package and opening up issues for each of your comments and suggestions.

After looking through the review issues, I found them all to be reasonable and valid, and encourage @rCarto to make the changes suggested. With regards to https://github.com/riatelab/osrm/issues/102, @JosiahParry I appreciate you raising some concerns about the size and completeness of the package. When `osrm` was submitted to JOSS it was re-circulated among the editorial team to see whether or not it was within scope during the pre-review phase as it was a "borderline" package in terms of size and scope. Due to its popularity and use across a number of different publications in different fields, we decided that the package would be considered "in-scope." However your point about feature completeness is very well taken -- @rCarto, I agree with @JosiahParry that it is very important for `osrm` to be fully feature complete with proper tests, input type-checking, and clear documentation with regards to expected input types, assumptions, and internal data transformations.
