### 2022-08-24 riatelab/osrm#104 @wcjochem
Typo `osrmTrip` example
A minor point, but I believe the final line of the example (https://github.com/riatelab/osrm/blob/master/R/osrmTrip.R#L38) should subset the points:

`plot(st_geometry(apotheke.sf[1:5,]), pch = 21, bg = "red", cex = 1, add = TRUE)`
