### 2022-08-24 riatelab/osrm#105 @wcjochem
osrmRoute argument order
When I was testing this package, I noticed that the order of the arguments for `osrmRoute` is different from `osrmTable`. Is there a reason for this?

This seems like a minor design question, but I was confused that there were 3 potential inputs. Could you consider having just `src` and `dst` as inputs, with `src` always required and `dst` be optional (or, if not supplied set it to be `src`)

For example:

```
osrmTable <- function(src, dst = src, exclude = NULL ...)
{
	# check inputs
	if(missing(src){ stop() }
}
```
