### 2022-08-24 riatelab/osrm#106 @wcjochem
Define 'isochrones'
In the readme, documentation and JOSS paper, please briefly define "isochrones" and "isometrics" to help users understand the outputs.
