### 2022-08-24 riatelab/osrm#107 @wcjochem
Difference between `osrmRoute` and `osrmTrip`
What is the difference between `osrmRoute` and `osrmTrip`? The documentation states that `osrmRoute` is for distances between two points, but the `loc` option allows for multiple stops along a route (as given in its example). `osrmTrip` also accepts multiple points. These two functions give different results for the same sets of points, so when should each function be used? Please clarify the function documentation and consider adding an example to demonstrate how to select the function.
