### 2022-08-24 riatelab/osrm#108 @wcjochem
Distance between multiple pairs of points?
What is the best way to get the shortest distance between multiple pairs of points? Please consider adding this as a usage example.

I tried `osrmRoute` with src and dst, but that only gives the results for the first rows. Perhaps there should be a warning to users about this, or add to the documentation.

When I tried `osrmTable`, I get the matrix of all pairwise comparisons.

`osrmRoute(src = apotheke.sf, dst = apotheke.df[sample(1:nrow(apotheke.df)),], returnclass="sf")`
`osrmTable(src = apotheke.sf[1:10,], dst = apotheke.df[11:20,])`
