### 2022-08-24 #4574 @wcjochem
Thank you @elbeejay for the opportunity to review this package.

I echo many of the questions/comments raised by @JosiahParry and @mikemahoney218. Therefore I have added only a few additional issues to the repository which were not covered by the other reviewers:
riatelab/osrm#104, riatelab/osrm#105, riatelab/osrm#106, riatelab/osrm#107, riatelab/osrm#108.

Many of my concerns are a combination of making the package clearer to users, through documentation/examples as well as clarifying the functionality.
