### 2022-09-01 aa7608b
docs: typo
fix #104
