### 2022-09-02 4ed48c1
docs: paper
address #92 comments on state of the field and phrasing suggestions
