+++ paper.bib
@@ -381,3 +381,11 @@ keywords = {Children, Distance analysis, Green spaces, Nature, Planning strategi
     note = {R package version 0.1.0},
     url = {https://CRAN.R-project.org/package=valhallr},
   }
+
+@Manual{osrmr,
+    title = {osrmr: Wrapper for the 'OSRM' API},
+    author = {Adrian Staempfli and Christoph Strauss},
+    year = {2021},
+    note = {R package version 0.1.36},
+    url = {https://CRAN.R-project.org/package=osrmr},
+  }
