### 2022-09-02 8086720
docs: add instructions on how to contribute
fix #103
