+++ man/osrmRoute.Rd
@@ -17,13 +17,13 @@ osrmRoute(
 }
 \arguments{
 \item{src}{a vector of identifier, longitude and latitude (WGS84), a vector 
-of longitude and latitude (WGS84) or an sf object of the origine point.}
+of longitude and latitude (WGS84) or an sf object of the origin point.}
 
 \item{dst}{a vector of identifier, longitude and latitude (WGS84), a vector 
 of longitude and latitude (WGS84) or an sf object of the destination point.}
 
 \item{loc}{a data.frame of identifier, longitude and latitude (WGS84) 
-or an sf object of via points. The first row is the origine, the last row 
+or an sf object of via points. The first row is the origin, the last row 
 is the destination.}
 
 \item{overview}{"full", "simplified" or FALSE. Use "full" to return the 
@@ -41,7 +41,7 @@ FALSE to return only time and distance.}
 \value{
 An sf LINESTRING is returned. \cr
 The sf LINESTRING contains 4 fields: identifiers of 
-origine and destination, travel time in minutes and travel distance in 
+origin and destination, travel time in minutes and travel distance in 
 kilometers.\cr\cr
 If overview is FALSE, a named numeric vector is returned. It contains travel 
 time (in minutes) and travel distance (in kilometers).
