### 2022-09-09 106ab9d
fix: modify input for osrmRoute() (vector of coordinates, data.frame, matrix, sfc POINTS, sf POINTS), modify documentation on WGS84 use
fix #94, fix #97
