+++ R/osrmRoute.R
@@ -1,14 +1,44 @@
 #' @name osrmRoute
 #' @title Get the Shortest Path Between Two Points
 #' @description Build and send an OSRM API query to get the travel geometry 
-#' between two points. This function interfaces the \emph{route} OSRM service. 
-#' @param src a vector of identifier, longitude and latitude (WGS84), a vector 
-#' of longitude and latitude (WGS84) or an sf object of the origin point.
-#' @param dst a vector of identifier, longitude and latitude (WGS84), a vector 
-#' of longitude and latitude (WGS84) or an sf object of the destination point.
-#' @param loc a data.frame of identifier, longitude and latitude (WGS84) 
-#' or an sf object of via points. The first row is the origin, the last row 
-#' is the destination.
+#' between two points. This function interfaces with the \emph{route} OSRM 
+#' service.\cr
+#' Use \code{src} and \code{dst} to get the shortest direct route between
+#' two points.\cr
+#' Use \code{loc} to get the shortest route between two points using
+#' waypoints. 
+#' 
+#' 
+#' @param src starting point of the route. 
+#' \code{src} can be: \itemize{
+#'   \item a vector of coordinates (longitude and latitude, WGS 84), 
+#'   \item a data.frame of longitudes and latitudes (WGS 84),
+#'   \item a matrix of longitudes and latitudes (WGS 84),
+#'   \item an sfc object of type POINT,
+#'   \item an sf object of type POINT.
+#'}
+#' If \code{src} is a data.frame, a matrix, an sfc object or an sf object then 
+#' only the first row or element is considered.
+#' @param dst destination of the route. 
+#' \code{dst} can be: \itemize{
+#'   \item a vector of coordinates (longitude and latitude, WGS 84), 
+#'   \item a data.frame of longitudes and latitudes (WGS 84),
+#'   \item a matrix of longitudes and latitudes (WGS 84),
+#'   \item an sfc object of type POINT,
+#'   \item an sf object of type POINT.
+#'}
+#' If \code{dst} is a data.frame, a matrix, an sfc object or an sf object then 
+#' only the first row or element is considered.
+#' 
+#' @param loc starting point, waypoints (optional) and destination of the 
+#' route. \code{loc} can be: \itemize{
+#'   \item a data.frame of longitudes and latitudes (WGS 84),
+#'   \item a matrix of longitudes and latitudes (WGS 84),
+#'   \item an sfc object of type POINT,
+#'   \item an sf object of type POINT.
+#'}
+#' The first row or element is the starting point, the last row or element is 
+#' the destination.
 #' @param overview "full", "simplified" or FALSE. Use "full" to return the 
 #' detailed geometry, use "simplified" to return a simplified geometry, use 
 #' FALSE to return only time and distance.
@@ -16,11 +46,21 @@
 #' @param osrm.server the base URL of the routing server.
 #' @param osrm.profile the routing profile to use, e.g. "car", "bike" or "foot".
 #' @param returnclass deprecated.
+
 #' @return
-#' An sf LINESTRING is returned. \cr
-#' The sf LINESTRING contains 4 fields: identifiers of 
-#' origin and destination, travel time in minutes and travel distance in 
-#' kilometers.\cr\cr
+#' The output of this function is an sf LINESTRING of the shortest route.\cr
+#' It contains 4 fields: \itemize{
+#'   \item starting point identifier (src row name or "src" if src is a vector 
+#'   of coordinates or an sfc object)
+#'   \item destination identifier (dst row name or "dst" if dst is a vector 
+#'   of coordinates or an sfc object)
+#'   \item travel time in minutes
[…]
