### 2022-09-09 faf8b31
fix: modify osrmTable to accept sfc object, better input type checking
fix #95
