+++ R/osrmTable.R
@@ -2,17 +2,29 @@
 #' @title Get Travel Time Matrices Between Points
 #' @description Build and send OSRM API queries to get travel time matrices 
 #' between points. This function interfaces the \emph{table} OSRM service. 
-#' @param loc a data frame containing 3 fields: points identifiers, longitudes 
-#' and latitudes (WGS84). It can also be an sf object. If so, row names are used 
-#' as identifiers.
-#' If loc parameter is used, all pair-wise distances are computed.
-#' @param src a data frame containing origin points identifiers, longitudes 
-#' and latitudes (WGS84). It can also be an sf object. If so, row names are used
-#' as identifiers. 
-#' If dst and src parameters are used, only pairs between scr/dst are computed.
-#' @param dst a data frame containing destination points identifiers, longitudes 
-#' and latitudes (WGS84). It can also be an sf object. If so, row names are used 
-#' as identifiers. 
+#' Use \code{src} and \code{dst} to set different origins and destinations.\cr
+#' Use \code{loc} to compute travel times or travel distances between all 
+#' points. 
+#' @param src origin points. 
+#' \code{src} can be: \itemize{
+#'   \item a data.frame of longitudes and latitudes (WGS 84),
+#'   \item a matrix of longitudes and latitudes (WGS 84),
+#'   \item an sfc object of type POINT,
+#'   \item an sf object of type POINT.
+#'}
+#' @param dst destination. 
+#' \code{dst} can be: \itemize{
+#'   \item a data.frame of longitudes and latitudes (WGS 84),
+#'   \item a matrix of longitudes and latitudes (WGS 84),
+#'   \item an sfc object of type POINT,
+#'   \item an sf object of type POINT.
+#'}
+#' @param loc points. \code{loc} can be: \itemize{
+#'   \item a data.frame of longitudes and latitudes (WGS 84),
+#'   \item a matrix of longitudes and latitudes (WGS 84),
+#'   \item an sfc object of type POINT,
+#'   \item an sf object of type POINT.
+#'}
 #' @param measure a character indicating what measures are calculated. It can 
 #' be "duration" (in minutes), "distance" (meters), or both c('duration',
 #' 'distance').  
@@ -25,8 +37,8 @@
 #' sources and destinations are the coordinates of 
 #' the origin and destination points actually used to compute the travel 
 #' times (WGS84).
-#' @details If loc, src or dst are data frames we assume that the 3 first 
-#' columns of the data.frame are: identifiers, longitudes and latitudes.\cr
+#' @details If loc, src or dst are data.frames or matrices it must have two
+#' colums containing longitudes and latitudes in WGS 84.\cr
 #' The OSRM demo server does not allow large queries (more than 10000 distances 
 #' or durations). 
 #' @note 
@@ -62,42 +74,35 @@
 #' distA4$durations[1:5,1:5]
 #' }
 #' @export
-osrmTable <- function(loc, src, dst, exclude, measure = "duration", 
+osrmTable <- function(src, 
+                      dst = src, 
+                      loc, 
+                      exclude, 
+                      measure = "duration", 
                       osrm.server = getOption("osrm.server"),
                       osrm.profile = getOption("osrm.profile")){
-  
-
   
   opt <- options(error = NULL)
   on.exit(options(opt), add=TRUE)
   
   url <- base_url(osrm.server, osrm.profile, "table")
-  
+
   # input management
   if(!missing(loc)){
-    if (methods::is(loc,"sf")){
-      loc <- sf_2_df(loc)
-    }
-    names(loc) <- c("id", "lon", "lat")
-    dst <- src <- loc
+    loc <- input_table(x = loc, id = 'loc')
+    dst_r <- src_r <- loc
[…]
