+++ R/osrmRoute.R
@@ -6,7 +6,7 @@
 #' Use \code{src} and \code{dst} to get the shortest direct route between
 #' two points.\cr
 #' Use \code{loc} to get the shortest route between two points using
-#' waypoints. 
+#' ordered waypoints. 
 #' 
 #' 
 #' @param src starting point of the route. 
