### 2022-09-09 cd45ea9
docs: update doc on osrmTable()
