### 2022-09-09 795f0f8
fix: modify osrmTrip() to accept sfc, check input objects, update doc
fix #107, fix #100, fix #99, fix #98
