+++ R/osrmRoute.R
@@ -17,6 +17,7 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
+#' If relevant row names are used as identifiers.\cr
 #' If \code{src} is a data.frame, a matrix, an sfc object or an sf object then 
 #' only the first row or element is considered.
 #' @param dst destination of the route. 
@@ -27,6 +28,7 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
+#' If relevant row names are used as identifiers.\cr
 #' If \code{dst} is a data.frame, a matrix, an sfc object or an sf object then 
 #' only the first row or element is considered.
 #' 
@@ -37,8 +39,10 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
-#' The first row or element is the starting point, the last row or element is 
-#' the destination.
+#' The first row or element is the starting point, the waypoints are used in 
+#' the order they are stored in \code{loc}, the last row or element is 
+#' the destination.\cr
+#' If relevant row names are used as identifiers.\cr
 #' @param overview "full", "simplified" or FALSE. Use "full" to return the 
 #' detailed geometry, use "simplified" to return a simplified geometry, use 
 #' FALSE to return only time and distance.
@@ -50,10 +54,8 @@
 #' @return
 #' The output of this function is an sf LINESTRING of the shortest route.\cr
 #' It contains 4 fields: \itemize{
-#'   \item starting point identifier (src row name or "src" if src is a vector 
-#'   of coordinates or an sfc object)
-#'   \item destination identifier (dst row name or "dst" if dst is a vector 
-#'   of coordinates or an sfc object)
+#'   \item starting point identifier
+#'   \item destination identifier
 #'   \item travel time in minutes
 #'   \item travel distance in kilometers.
 #'   }

[…]
