### 2022-09-09 riatelab/osrm#105 @rCarto
fixed in faf8b311ab0f69917a99c3384b3b1f23fe22310c
`osrmRoute()` and  `osrmTable()` arguments are now aligned.
I'll keep the loc arg for backward compatibility and ease of use.
