### 2022-09-09 riatelab/osrm#96 @rCarto
> One of the undocumented features of `osrm` is that projected coordinate systems are first transformed to wgs84. The documentation states that geometry _must_ be lon lat. This is false.

I understand that the documentation was not clear on the subject. WGS84 is only needed for non-geographic input formats.
When using spatial object the transformation to WGS84 is invisible to the user and the results are returned in the original CRS.
I'm not sure why we should warn the user about this transformation.

I've modified the doc to make it more clear.
