### 2022-09-09 riatelab/osrm#108 @rCarto
This operation is not possible with `osrm` functions only.
On could use a for loop for example:
``` r
library(osrm)
#> Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
#> Routing: OSRM - http://project-osrm.org/
library(sf)
#> Linking to GEOS 3.9.0, GDAL 3.2.2, PROJ 7.2.1; sf_use_s2() is TRUE
options(osrm.server = "http://0.0.0.0:5000/", osrm.profile = "driving")
ap <- st_read(system.file("gpkg/apotheke.gpkg", package = "osrm"),
              quiet = TRUE)
routes_list <- vector("list", 10)
for (i in seq_along(routes_list)){
  routes_list[[i]] <- osrmRoute(src = ap[1, ], dst = ap[1+i, ])
  # ADD THIS WHEN USING DEMO SERVER
  # Sys.sleep(1)
}
routes <- do.call(rbind, routes_list)
plot(st_geometry(routes))
```

![](https://i.imgur.com/slou5hT.png)

> I tried `osrmRoute` with src and dst, but that only gives the results for the first rows. Perhaps there should be a warning to users about this, or add to the documentation.

I've added a message and updated the documentation on that subject with 106ab9d19d8266559c1164ae2fdf51f2651755e9
