+++ R/osrmRoute.R
@@ -17,7 +17,7 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
-#' If relevant row names are used as identifiers.\cr
+#' If relevant, row names are used as identifiers.\cr
 #' If \code{src} is a data.frame, a matrix, an sfc object or an sf object then 
 #' only the first row or element is considered.
 #' @param dst destination of the route. 
@@ -28,7 +28,7 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
-#' If relevant row names are used as identifiers.\cr
+#' If relevant, row names are used as identifiers.\cr
 #' If \code{dst} is a data.frame, a matrix, an sfc object or an sf object then 
 #' only the first row or element is considered.
 #' 
@@ -42,7 +42,7 @@
 #' The first row or element is the starting point, the waypoints are used in 
 #' the order they are stored in \code{loc}, the last row or element is 
 #' the destination.\cr
-#' If relevant row names are used as identifiers.\cr
+#' If relevant, row names are used as identifiers.\cr
 #' @param overview "full", "simplified" or FALSE. Use "full" to return the 
 #' detailed geometry, use "simplified" to return a simplified geometry, use 
 #' FALSE to return only time and distance.
