+++ man/osrmRoute.Rd
@@ -24,7 +24,7 @@ osrmRoute(
   \item an sfc object of type POINT,
   \item an sf object of type POINT.
 }
-If relevant row names are used as identifiers.\cr
+If relevant, row names are used as identifiers.\cr
 If \code{src} is a data.frame, a matrix, an sfc object or an sf object then 
 only the first row or element is considered.}
 
@@ -36,7 +36,7 @@ only the first row or element is considered.}
   \item an sfc object of type POINT,
   \item an sf object of type POINT.
 }
-If relevant row names are used as identifiers.\cr
+If relevant, row names are used as identifiers.\cr
 If \code{dst} is a data.frame, a matrix, an sfc object or an sf object then 
 only the first row or element is considered.}
 
@@ -50,7 +50,7 @@ route. \code{loc} can be: \itemize{
 The first row or element is the starting point, the waypoints are used in 
 the order they are stored in \code{loc}, the last row or element is 
 the destination.\cr
-If relevant row names are used as identifiers.\cr}
+If relevant, row names are used as identifiers.\cr}
 
 \item{overview}{"full", "simplified" or FALSE. Use "full" to return the 
 detailed geometry, use "simplified" to return a simplified geometry, use 
