### 2022-09-12 f7c874f
fix: correct conditional test + typos
