### 2022-09-12 b45c462
test: add test for internal function (utils.R), update all test except osrmIso*
fix #93
