### 2022-09-12 8306fc9
test: use more permissive tests for sf objects (testing machines have != GDAL & PROJ versions)
