+++ inst/tinytest/test_internals_input_route.R
@@ -102,11 +102,11 @@ expect_identical(osrm:::input_route(x =  x_m[1,,drop = F],
 #                                         single = TRUE,
 #                                         all.ids = FALSE)
 # saveRDS(input_route_out_sfc, "inst/tinytest/input_route_out_sfc.rds")
-expect_identical(osrm:::input_route(x = x_sfc[1], 
-                                    id = "src", 
-                                    single = TRUE, 
-                                    all.ids = FALSE), 
-                 readRDS("input_route_out_sfc.rds"))
+expect_equivalent(osrm:::input_route(x = x_sfc[1], 
+                                     id = "src", 
+                                     single = TRUE, 
+                                     all.ids = FALSE), 
+                  readRDS("input_route_out_sfc.rds"))
 
 # input sf
 # input_route_out_sf <- osrm:::input_route(x = x_sf[1,],
@@ -114,11 +114,11 @@ expect_identical(osrm:::input_route(x = x_sfc[1],
 #                                         single = TRUE,
 #                                         all.ids = FALSE)
 # saveRDS(input_route_out_sf, "inst/tinytest/input_route_out_sf.rds")
-expect_identical(osrm:::input_route(x = x_sf[1, ], 
-                                    id = "src", 
-                                    single = TRUE, 
-                                    all.ids = FALSE), 
-                 readRDS("input_route_out_sf.rds"))
+expect_equivalent(osrm:::input_route(x = x_sf[1, ], 
+                                     id = "src", 
+                                     single = TRUE, 
+                                     all.ids = FALSE), 
+                  readRDS("input_route_out_sf.rds"))
 
 
 ######## MULTI
@@ -152,11 +152,11 @@ expect_identical(osrm:::input_route(x =  x_m[1:4,],
 #                                         single = FALSE,
 #                                         all.ids = FALSE)
 # saveRDS(input_route_out_sfc_m, "inst/tinytest/input_route_out_sfc_m.rds")
-expect_identical(osrm:::input_route(x = x_sfc[1:4], 
-                                    id = "loc", 
-                                    single = FALSE, 
-                                    all.ids = FALSE), 
-                 readRDS("input_route_out_sfc_m.rds"))
+expect_equivalent(osrm:::input_route(x = x_sfc[1:4], 
+                                     id = "loc", 
+                                     single = FALSE, 
+                                     all.ids = FALSE), 
+                  readRDS("input_route_out_sfc_m.rds"))
 
 # input sf
 # input_route_out_sf_m <- osrm:::input_route(x = x_sf[1:4,],
@@ -164,11 +164,11 @@ expect_identical(osrm:::input_route(x = x_sfc[1:4],
 #                                         single = FALSE,
 #                                         all.ids = FALSE)
 # saveRDS(input_route_out_sf_m, "inst/tinytest/input_route_out_sf_m.rds")
-expect_identical(osrm:::input_route(x = x_sf[1:4, ], 
-                                    id = "loc", 
-                                    single = FALSE, 
-                                    all.ids = FALSE), 
-                 readRDS("input_route_out_sf_m.rds"))
+expect_equivalent(osrm:::input_route(x = x_sf[1:4, ], 
+                                     id = "loc", 
+                                     single = FALSE, 
+                                     all.ids = FALSE), 
+                  readRDS("input_route_out_sf_m.rds"))
 
 
 
@@ -203,11 +203,11 @@ expect_identical(osrm:::input_route(x =  x_m[1:4,],
 #                                         single = FALSE,
 #                                         all.ids = TRUE)
 # saveRDS(input_route_out_sfc_m_id, "inst/tinytest/input_route_out_sfc_m_id.rds")
[…]
