+++ inst/tinytest/test_internals_rgrid.R
@@ -5,4 +5,4 @@ res <- 30
 # rgrid_out <- osrm:::rgrid(loc, dmax, res)
 # saveRDS(rgrid_out, 'inst/tinytest/rgrid_out.rds')
 rgrid_out <- osrm:::rgrid(loc, dmax, res)
-expect_identical(rgrid_out, readRDS("rgrid_out.rds"))
+expect_equivalent(rgrid_out, readRDS("rgrid_out.rds"))
