### 2022-09-13 b9c1381
test: recompute CRS from test suite
