### 2022-09-13 232f144
refactor: use internal function for osrmIso*() funs
