+++ R/osrmIsometric.R
@@ -1,7 +1,7 @@
-#' @name osrmIsochrone
-#' @title Get Polygons of Isochrones
+#' @name osrmIsometric
+#' @title Get Polygons of Isodistances
 #' @description This function computes areas that are reachable within a 
-#' given time span from a point and returns the reachable regions as 
+#' given road distance from a point and returns the reachable regions as 
 #' polygons. 
 #' @param loc origin point. \code{loc} can be: \itemize{
 #'   \item a vector of coordinates (longitude and latitude, WGS 84), 
@@ -12,11 +12,11 @@
 #'}
 #' If \code{loc} is a data.frame, a matrix, an sfc object or an sf object then 
 #' only the first row or element is considered.
-#' @param breaks a numeric vector of isochrone values (in minutes).
+#' @param breaks a numeric vector of isometric values (in meters).
 #' @param exclude pass an optional "exclude" request option to the OSRM API.
-#' @param res number of points used to compute isochrones, one side of the square
-#' grid, the total number of points will be res*res. Increase res to obtain more
-#' detailed isochrones. 
+#' @param res number of points used to compute isodistances, one side of the 
+#' square grid, the total number of points will be res*res. Increase res to 
+#' obtain more detailed isodistances. 
 #' @param returnclass deprecated.
 #' @param osrm.server the base URL of the routing server.
 #' getOption("osrm.server") by default.
@@ -24,19 +24,17 @@
 #' (when using the routing.openstreetmap.de test server).
 #' getOption("osrm.profile") by default.
 #' @return
-#' The output of this function is an sf MULTIPOLYGON of isochrones.\cr
+#' The output of this function is an sf MULTIPOLYGON of isodistances.\cr
 #' It contains 3 fields: \itemize{
 #'   \item id, an identifier
-#'   \item isomin, the minimum value of the isochrone polygon in minutes
-#'   \item isomax, the maximum value of the isochrone polygon in minutes
+#'   \item isomin, the minimum value of the isodistance polygon in meters
+#'   \item isomax, the maximum value of the isodistance polygon in meters
 #' }
 #' If loc is a vector, a data.frame or a matrix the coordinate 
 #' reference system (CRS) of the output is EPSG:4326 (WGS84).\cr 
 #' If loc is an sfc or sf object, the output has the same CRS 
 #' as loc.\cr
-#' @importFrom sf st_as_sf st_crs st_transform st_convex_hull st_union 
-#' st_intersects st_bbox st_buffer st_distance st_make_grid
-#' @importFrom mapiso mapiso
+#' @importFrom sf st_as_sf st_crs st_transform st_convex_hull st_union st_intersects
 #' @export
 #' @examples
 #' \dontrun{
@@ -44,12 +42,12 @@
 #' apotheke.sf <- st_read(system.file("gpkg/apotheke.gpkg", package = "osrm"),
 #'                        quiet = TRUE)
 #' # Get isochones with lon/lat coordinates
-#' iso <- osrmIsochrone(loc = c(13.43,52.47), breaks = seq(0,12,2))
+#' iso <- osrmIsometric(loc = c(13.43,52.47), breaks = seq(0,500,100))
 #' # Map
 #' plot(iso["isomax"], breaks = sort(unique(c(iso$isomin, iso$isomax))))
 #' 
 #' # Get isochones with an sf POINT
-#' iso2 <- osrmIsochrone(loc = apotheke.sf[11,], breaks = seq(0,12,2))
+#' iso2 <- osrmIsometric(loc = apotheke.sf[11,], breaks = seq(0,500,100))
 #' # Map
[…]
