+++ inst/tinytest/test_osrmIsometric.R
@@ -3,13 +3,13 @@ if(demo_server){
   options(osrm.server = "https://routing.openstreetmap.de/", 
           osrm.profile = "car")
   wait()
-  r <- osrmIsochrone(loc = c(13.43,52.47), breaks = seq(0,14,2), res = 10 )
+  r <- osrmIsometric(loc = c(13.43,52.47), breaks = seq(0,500,100), res = 10 )
   expect_true(inherits(r, "sf"))
   expect_identical(st_crs(r), st_crs("EPSG:4326"))
   expect_identical(colnames(r), 
                    c("id", "isomin", "isomax", "geometry"))
   wait()
-  r <- osrmIsochrone(loc = x_sf[10, ], breaks = seq(0,14,2), res = 10 )
+  r <- osrmIsometric(loc = x_sf[11, ], breaks = seq(0,500,100), res = 10 )
   expect_true(inherits(r, "sf"))
   expect_identical(st_crs(r), st_crs(x_sf))
   expect_identical(colnames(r), 
@@ -18,13 +18,13 @@ if(demo_server){
   
   ################# DEMO BIKE #####################
   options(osrm.server = "https://routing.openstreetmap.de/", osrm.profile = "bike")
-  r <- osrmIsochrone(loc = c(13.43,52.47), breaks = seq(0,14,2), res = 10 )
+  r <- osrmIsometric(loc = c(13.43,52.47), breaks = seq(0,500,100), res = 10 )
   expect_true(inherits(r, "sf"))
   expect_identical(st_crs(r), st_crs("EPSG:4326"))
   expect_identical(colnames(r), 
                    c("id", "isomin", "isomax", "geometry"))
   wait()
-  r <- osrmIsochrone(loc = x_sf[10, ], breaks = seq(0,14,2), res = 10 )
+  r <- osrmIsometric(loc = x_sf[11, ], breaks = seq(0,500,100), res = 10 )
   expect_true(inherits(r, "sf"))
   expect_identical(st_crs(r), st_crs(x_sf))
   expect_identical(colnames(r), 
@@ -33,13 +33,13 @@ if(demo_server){
   
   ############## DEMO FOOT #################"""""
   options(osrm.server = "https://routing.openstreetmap.de/", osrm.profile = "foot")
-  r <- osrmIsochrone(loc = c(13.43,52.47), breaks = seq(0,14,2), res = 10 )
+  r <- osrmIsometric(loc = c(13.43,52.47), breaks = seq(0,500,100), res = 10 )
   expect_true(inherits(r, "sf"))
   expect_identical(st_crs(r), st_crs("EPSG:4326"))
   expect_identical(colnames(r), 
                    c("id", "isomin", "isomax", "geometry"))
   wait()
-  r <- osrmIsochrone(loc = x_sf[10, ], breaks = seq(0,14,2), res = 10 )
+  r <- osrmIsometric(loc = x_sf[11, ], breaks = seq(0,500,100), res = 10 )
   expect_true(inherits(r, "sf"))
   expect_identical(st_crs(r), st_crs(x_sf))
   expect_identical(colnames(r), 
@@ -48,7 +48,7 @@ if(demo_server){
   
   ############# server param ##################""
   wait()
-  r <- osrmIsochrone(loc = c(13.43,52.47), breaks = seq(0,14,2), res = 10, 
+  r <- osrmIsometric(loc = c(13.43,52.47), breaks = seq(0,500,100), res = 10,
                      osrm.server = "https://router.project-osrm.org/", 
                      osrm.profile = "driving")
   expect_true(inherits(r, "sf"))
@@ -58,12 +58,12 @@ if(demo_server){
   
   # server error
[…]
