### 2022-09-13 b2b2263
docs: better documentation and examples
