+++ R/osrmTable.R
@@ -12,6 +12,7 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
+#' If relevant, row names are used as identifiers.
 #' @param dst destination. 
 #' \code{dst} can be: \itemize{
 #'   \item a data.frame of longitudes and latitudes (WGS 84),
@@ -19,12 +20,14 @@
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
+#' If relevant, row names are used as identifiers.
 #' @param loc points. \code{loc} can be: \itemize{
 #'   \item a data.frame of longitudes and latitudes (WGS 84),
 #'   \item a matrix of longitudes and latitudes (WGS 84),
 #'   \item an sfc object of type POINT,
 #'   \item an sf object of type POINT.
 #'}
+#' If relevant, row names are used as identifiers.
 #' @param measure a character indicating what measures are calculated. It can 
 #' be "duration" (in minutes), "distance" (meters), or both c('duration',
 #' 'distance').  
@@ -32,31 +35,35 @@
 #' (not allowed with the OSRM demo server).
 #' @param osrm.server the base URL of the routing server.
 #' @param osrm.profile the routing profile to use, e.g. "car", "bike" or "foot".
-#' @return A list containing 3 data frames is returned. 
-#' durations is the matrix of travel times (in minutes), 
-#' sources and destinations are the coordinates of 
-#' the origin and destination points actually used to compute the travel 
-#' times (WGS84).
-#' @details If loc, src or dst are data.frames or matrices it must have two
-#' colums containing longitudes and latitudes in WGS 84.\cr
-#' The OSRM demo server does not allow large queries (more than 10000 distances 
-#' or durations). 
+#' @return
+#' The output of this function is a list composed of one or two matrices 
+#' and 2 data.frames 
+#' \itemize{
+#'   \item{durations}: a matrix of travel times (in minutes)
+#'   \item{distances}: a matrix of distances (in meters)
+#'   \item{sources}: a data.frame of the coordinates of the points actually
+#'   used as starting points (EPSG:4326 - WGS84)
+#'   \item{sources}: a data.frame of the coordinates of the points actually
+#'   used as destinations (EPSG:4326 - WGS84)
+#'   }
 #' @note 
-#' If you want to get a large number of distances make sure to set the 
-#' "max-table-size" option (Max. locations supported in table) of the OSRM 
-#' server accordingly.
+#' The OSRM demo server does not allow large queries (more than 10000 distances 
+#' or durations).\cr
+#' If you use your own server and if you want to get a large number of distances
+#' make sure to set the "max-table-size" option (Max. locations supported in 
+#' table) of the OSRM server accordingly.
 #' @examples
 #' \dontrun{
 #' # Inputs are data frames
 #' apotheke.df <- read.csv(system.file("csv/apotheke.csv", package = "osrm"))
 #' # Travel time matrix
-#' distA <- osrmTable(loc = apotheke.df[1:50, c("id","lon","lat")])
+#' distA <- osrmTable(loc = apotheke.df[1:50, c("lon","lat")])
 #' # First 5 rows and columns
 #' distA$durations[1:5,1:5]
 #' 
 #' # Travel time matrix with different sets of origins and destinations
-#' distA2 <- osrmTable(src = apotheke.df[1:10,c("id","lon","lat")],
-#'                     dst = apotheke.df[11:20,c("id","lon","lat")])
[…]
