### 2022-09-13 07dc4d7
test: internal grid input/output test
