+++ DESCRIPTION
@@ -19,7 +19,6 @@ Imports:
     curl,
     utils, 
     mapiso,
-    methods,
     googlePolylines,
     sf
 Depends:
