### 2022-09-14 f36785f
docs: add feature description and example in paper
