+++ paper.md
@@ -60,9 +60,142 @@ Several packages exist to compute routes, trips or distance matrices. Most of th
 Among these packages, `osrm` has the advantage of using OSRM, which is easy to install and run on a local or remote server, of giving access to most OSRM services and of handling spatial data formats. 
 
 # Features 
-This is a short overview of the main features of `osrm`. The dataset used here is shipped with the package, it is a sample of 100 random pharmacies in Berlin ([© OpenStreetMap contributors](https://www.openstreetmap.org/copyright/en) stored in a [geopackage](https://www.geopackage.org/) file. 
+
+This is a short overview of the main features of `osrm`. The dataset used here is shipped with the package, it is a sample of 100 random pharmacies in Berlin ([© OpenStreetMap contributors](https://www.openstreetmap.org/copyright/en)) stored in a [geopackage](https://www.geopackage.org/) file.  
+
+* `osrmTable()` gives access to the *table* OSRM service. In this example we use this function to get the median time needed to access any pharmacy from any other pharmacy.   
+
+``` r
+library(osrm)
+```
+
+    ## Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
+
+    ## Routing: OSRM - http://project-osrm.org/
 
 
+``` r
+library(sf)
+```
+
+    ## Linking to GEOS 3.9.0, GDAL 3.2.2, PROJ 7.2.1; sf_use_s2() is TRUE
+
+``` r
+pharmacy <- st_read(system.file("gpkg/apotheke.gpkg", package = "osrm"), 
+                    quiet = TRUE)
+travel_time <- osrmTable(loc = pharmacy)
+travel_time$durations[1:5,1:5]
+```
+
+    ##      1    2    3    4    5
+    ## 1  0.0 21.1 33.4 21.2 12.6
+    ## 2 22.1  0.0 42.3 16.1 20.2
+    ## 3 33.0 43.0  0.0 30.5 27.4
+    ## 4 20.1 15.3 29.7  0.0 12.7
+    ## 5 10.2 20.3 26.8 12.3  0.0
+
+``` r
+diag(travel_time$durations) <- NA
+median(travel_time$durations, na.rm = TRUE)
+```
+
+    ## [1] 21.4
+
+The median time needed to access any pharmacy from any other pharmacy is 21.4 minutes. 
+
+
+* `osrmRoute()` is used to compute the shortest route between two points. Here we compute the shortest route between the two first pharmacies. 
+
+``` r
+(route <- osrmRoute(src = pharmacy[1, ], pharmacy[2, ]))
+```
+
+    ## Simple feature collection with 1 feature and 4 fields
+    ## Geometry type: LINESTRING
+    ## Dimension:     XY
+    ## Bounding box:  xmin: -13170.51 ymin: 5837172 xmax: -3875.771 ymax: 5841047
+    ## Projected CRS: WGS 84 / UTM zone 34N
+    ##     src dst duration distance                       geometry
+    ## 1_2   1   2 21.11667   12.348 LINESTRING (-13170.51 58410...
+
+This route is 12.3 kilometers long and it takes 21.1 minutes to drive through it. 
+
+``` r
+plot(st_geometry(route))
+plot(st_geometry(pharmacy[1:2,]), pch = 20, add = T, cex = 1.5)
+```
+
+![](route.png)
+
+
[…]
