### 2022-09-14 3d41d24
docs: update README (from paper)
