+++ README.md
@@ -38,120 +38,165 @@ usage.](https://routing.openstreetmap.de/about.html)
 
 ## Features
 
--   `osrmTable` Get travel time matrices between points.
-
--   `osrmRoute` Get the shortest path between two points.
-
--   `osrmTrip` Get the travel geometry between multiple unordered
-    points.
-
--   `osrmIsochrone` Get polygons of isochrones.
+-   `osrmTable()` uses the *table* service to query time/distance
+    matrices,
+-   `osrmRoute()` uses the *route* service to query routes,
+-   `osrmTrip()` uses the *trip* service to query trips,
+-   `osrmIsochone()` and `osrmIsometric()` use multiple `osrmTable()`
+    calls to create isochrones or isometric polygons.
 
 ## Demo
 
-### `osrmTable()`
+This is a short overview of the main features of `osrm`. The dataset
+used here is shipped with the package, it is a sample of 100 random
+pharmacies in Berlin ([© OpenStreetMap
+contributors](https://www.openstreetmap.org/copyright/en)) stored in a
+[geopackage](https://www.geopackage.org/) file.
+
+-   `osrmTable()` gives access to the *table* OSRM service. In this
+    example we use this function to get the median time needed to access
+    any pharmacy from any other pharmacy.
 
 ``` r
 library(osrm)
 ```
 
+    ## Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
+
+    ## Routing: OSRM - http://project-osrm.org/
+
 ``` r
 library(sf)
-apotheke.sf <- st_read(system.file("gpkg/apotheke.gpkg", package = "osrm"), 
-                       quiet = TRUE)
-# Travel time matrix
-distA <- osrmTable(loc = apotheke.sf[1:5,])
-distA$durations
 ```
 
-<small>
-
-|     |    1 |    2 |    3 |    4 |    5 |
-|:----|-----:|-----:|-----:|-----:|-----:|
-| 1   |  0.0 | 21.1 | 33.4 | 21.2 | 12.6 |
-| 2   | 22.1 |  0.0 | 42.3 | 16.1 | 20.2 |
-| 3   | 33.0 | 43.0 |  0.0 | 30.5 | 27.4 |
-| 4   | 20.1 | 15.3 | 29.7 |  0.0 | 12.7 |
-| 5   | 10.2 | 20.3 | 26.8 | 12.3 |  0.0 |
-
-</small>
-
-### `osrmRoute()`
+    ## Linking to GEOS 3.9.0, GDAL 3.2.2, PROJ 7.2.1; sf_use_s2() is TRUE
 
 ``` r
-library(maptiles)
-library(mapsf)
-# Transform to webmercator for a better display of map tiles
-apotheke.sf <- st_transform(apotheke.sf, 3857)
-# Route
-route <- osrmRoute(src = apotheke.sf[74,], dst = apotheke.sf[55,],
-                   overview = "full")
-# Get map tiles
-osm <- get_tiles(x = route, crop = TRUE, zoom = 13)
-# Map
-theme <- mf_theme(mar = c(0,0,1.2,0), inner = FALSE, line = 1.2, cex = .9, 
-                  pos = "center", tab = FALSE)
-mf_export(osm,filename = "img/route.png", width = ncol(osm), theme = theme)
-mf_raster(osm, add = TRUE)
-mf_map(route, lwd = 4, add = TRUE, col = "blue")
-mf_map(route, lwd = 1, col = "white", add = TRUE)
-mf_map(apotheke.sf[c(74,55),], pch = 20, col = "red", add = TRUE)
[…]
