+++ README.Rmd
@@ -31,120 +31,148 @@ You can run your own instance of OSRM following guidelines provided [here](https
 
 ## Features
 
-* `osrmTable` Get travel time matrices between points.
-
-* `osrmRoute` Get the shortest path between two points.
-
-* `osrmTrip` Get the travel geometry between multiple unordered points.
-
-* `osrmIsochrone` Get polygons of isochrones.
-
+- `osrmTable()` uses the *table* service to query time/distance matrices,
+- `osrmRoute()` uses the *route* service to query routes,
+- `osrmTrip()` uses the *trip* service to query trips,
+- `osrmIsochone()` and `osrmIsometric()` use multiple `osrmTable()` calls to create isochrones or isometric polygons.
 
 ## Demo
 
-### `osrmTable()`
+This is a short overview of the main features of `osrm`. The dataset used here is shipped with the package, it is a sample of 100 random pharmacies in Berlin ([© OpenStreetMap contributors](https://www.openstreetmap.org/copyright/en)) stored in a [geopackage](https://www.geopackage.org/) file.  
 
+* `osrmTable()` gives access to the *table* OSRM service. In this example we use this function to get the median time needed to access any pharmacy from any other pharmacy.   
 
-```{r pkg, results='hide', message=FALSE}
+``` r
 library(osrm)
 ```
 
-```{r init, echo = FALSE, include=FALSE}
-options(osrm.server = "http://0.0.0.0:5000/", osrm.profile = "car")
-```
+    ## Data: (c) OpenStreetMap contributors, ODbL 1.0 - http://www.openstreetmap.org/copyright
+
+    ## Routing: OSRM - http://project-osrm.org/
 
 
-```{r table, results='hide', message=FALSE}
+``` r
 library(sf)
-apotheke.sf <- st_read(system.file("gpkg/apotheke.gpkg", package = "osrm"), 
-                       quiet = TRUE)
-# Travel time matrix
-distA <- osrmTable(loc = apotheke.sf[1:5,])
-distA$durations
 ```
 
-<small>
-```{r table_display, echo=FALSE}
-knitr::kable(distA$durations, row.names = TRUE)
-```
-</small>
+    ## Linking to GEOS 3.9.0, GDAL 3.2.2, PROJ 7.2.1; sf_use_s2() is TRUE
 
-### `osrmRoute()`
-
-```{r route, results='hide', message=FALSE}
-library(maptiles)
-library(mapsf)
-# Transform to webmercator for a better display of map tiles
-apotheke.sf <- st_transform(apotheke.sf, 3857)
-# Route
-route <- osrmRoute(src = apotheke.sf[74,], dst = apotheke.sf[55,],
-                   overview = "full")
-# Get map tiles
-osm <- get_tiles(x = route, crop = TRUE, zoom = 13)
-# Map
-theme <- mf_theme(mar = c(0,0,1.2,0), inner = FALSE, line = 1.2, cex = .9, 
-                  pos = "center", tab = FALSE)
-mf_export(osm,filename = "img/route.png", width = ncol(osm), theme = theme)
-mf_raster(osm, add = TRUE)
-mf_map(route, lwd = 4, add = TRUE, col = "blue")
-mf_map(route, lwd = 1, col = "white", add = TRUE)
-mf_map(apotheke.sf[c(74,55),], pch = 20, col = "red", add = TRUE)
-mf_title("osrmRoute()")
[…]
