### 2022-09-15 #4574 @rCarto
Hello @elbeejay , @wcjochem , @JosiahParry and @mikemahoney218

Here is the list of modifications I have applied to the package in order to take into account your comments:

* State of the field
riatelab/osrm#92
I've added a reference about `osrmr` pkg.

* How to use the package
riatelab/osrm#92
I've added a section on how to use the package with a short demonstration of key functionalities in the paper and in the README

* Tests
riatelab/osrm#93
I've extracted some processing steps in their own functions and I've added tests that check the output of internal functions based on fixed input and output files (see [test_internals_input_route.R](https://github.com/riatelab/osrm/blob/3d41d24f827ad45942dd766c9d4efd16791f24c2/inst/tinytest/test_internals_input_route.R) for example).

* Inputs management
riatelab/osrm#94
riatelab/osrm#95
riatelab/osrm#98
riatelab/osrm#99
I've modified the code to better test inputs. Now, most of the functions accept data.frame of coordinates, matrix of coordinates, sfc and sf objects. I've change the way to set identifiers, row.names (data.frame, matrix, sf obj) or indexes (sfc obj, unnamed matrices) are used.

* WGS84 requirement
riatelab/osrm#96
riatelab/osrm#97
The documentation was unclear about which input need to use lon/lat wgs84 coordinates. I've tried to modify the documentation to make it more clear.
If the input is a non-spatial object (vector of coordinates, data.frames and matrices) it must use WGS84 coordinates and the output use WGS84.
If the input is a spatial object (sf or sfc object) it can use any CRS and the output use the input CRS.

* `osrmTrip()` documentation
riatelab/osrm#100
The documentation of `osrmTrip()` has been updated to better explain its output.

* Contributing guidelines
riatelab/osrm#103
I've added a CONTRIBUTING.md file describing contributing guidelines. This file is referenced at the bottom of the README file along with some shorter guidelines.

* `osrmRoute()` argument order
riatelab/osrm#105
`osrmRoute()` and `osrmTable()` arguments are now aligned. I'll keep the loc arg for backward compatibility and ease of use.

* Define 'isochrones'
riatelab/osrm#106
Documentation has been updated to define isochrones.

* Difference between `osrmRoute()` and `osrmTrip()`
riatelab/osrm#107
I've updated the documentation to better explain both functions.

* Typos
riatelab/osrm#101
riatelab/osrm#104
riatelab/osrm#92
Typos are corrected
