### 2022-09-19 5244a44
feat: add osrmNearest(), interface with nearest service
