### 2022-09-19 64f2b27
fix: avoid generating hints, marginal performence improvement
