### 2022-10-07 #4574 @JosiahParry
Thanks for the reminder! I'll be funemployed for a week next week. I'll get
to it then.

On Thu, Oct 6, 2022 at 9:27 PM Michael Mahoney ***@***.***>
wrote:

> Hey @elbeejay <https://github.com/elbeejay> -- sorry for the delay, I'm
> currently not paying a ton of attention to GitHub or emails (through the
> 21st). I'll review the changes in the next week.
>
