### 2022-10-11 62ab3bf
docs: update documentation with thelast OSRM version tested
