+++ R/package.R
@@ -29,7 +29,7 @@
 #' 
 #' @note
 #' This package relies on the usage of a running OSRM service (tested with 
-#' version 5.26.0 of the OSRM API).\cr
+#' version 5.27.0 of the OSRM API).\cr
 #' 
 #' 
 #' To set the OSRM server, change the \code{osrm.server} option:\cr 
