### 2022-10-11 752a3ac
test: switch on demo server test suite
