### 2022-10-13 #4574 @elbeejay
Thanks for following up @wcjochem, would you mind checking off the remainder of the boxes in your [checklist above](https://github.com/openjournals/joss-reviews/issues/4574#issuecomment-1187925369) if those areas that previously needed some work have now been addressed?

Regarding your concern about the new dependency, I appreciate the note about your difficulties installing it on Windows. @rCarto have you experienced this? The CI pipelines do include a Windows build, so it seems like it *can* be done. If there are known difficulties @rCarto, I would suggest being more descriptive in the installation instructions to help users get a functional installation on Windows machines.
