### 2022-10-13 riatelab/osrm#110 @JosiahParry
Improve error handling osrmIsochrone(): out of bounds
When providing a number that is out of bounds for longitude and latitude an error message says the provided values are non-finite.
This should error with a more appropriate message like `"[lon/lat] is bounded by the interval [-180, 180]"`

```r
# provide informative error for out of bounds vector coordinates
osrmIsochrone(c(700, -999), breaks = seq(0, 12, 3))
#> Error in seq.default(from = xf - dmax, to = xf + dmax, length.out = res) :
#>   'from' must be a finite number
```
