### 2022-10-13 riatelab/osrm#112 @JosiahParry
Improve error handling `osrmIsochrone()`: handle empty isochrones
When providing a location for an isochrone for which there is (assumedly) no data, `osrmIsochrone()` errors out with a cryptic message.

I think it would be useful if 1) the error was converted to a warning and 2) the function returned an empty sf object.

```r
 osrmIsochrone(c(10, 20), breaks = seq(0, 12, 3))
#> Error in if (any(idx)) { : missing value where TRUE/FALSE needed
```

Assuming `idx` is a logical vector, you could write something like:

```r
if (length(idx) == 0) {
  empty_res <- st_sf(
    crs = 4326,
    id = integer(),
    isomin = numeric(),
    isomax = numeric(),
    geometry = st_sfc()
  )

  return(empty_res)
}
#> Simple feature collection with 0 features and 3 fields
#> Bounding box:  xmin: NA ymin: NA xmax: NA ymax: NA
#> Geodetic CRS:  WGS 84
#> [1] id       isomin   isomax   geometry
#> <0 rows> (or 0-length row.names)
```

At minimum, if errors are preferred, an improved error message would be useful for users.
