### 2022-10-13 riatelab/osrm#113 @JosiahParry
Improve documentation `osrmIsochrone()`: `breaks` argument
The `breaks` arugment of `osrmIsocrhone()` says `"a numeric vector of isochrone values (in minutes)."`  This language requires a bit of specialized knowledge. A google search for "isochrome values" does not resolve any confusions. My suspicion is that the `breaks` are intervals in the travel distance in units of minutes. But this is just a hunch.

Please document `breaks` a bit more clearly either in the argument itself or in the description. Per #106 I do not see a definition of "isochrone" in `osrmIsochrone()`, the readme, or any other functions man docs.
