### 2022-10-13 riatelab/osrm#106 @JosiahParry
I do not think this is resolved. I cannot find a definition of isochrone in the readme or various man pages. Though, I think the readme gets close at https://github.com/riatelab/osrm/blob/752a3ac9500f4f1895c4a137486a0fc31439841c/README.Rmd#L149.
