### 2022-10-13 279a044
fix: improve error handling for vector of coordinate input
fix #111

[ci skip]
