### 2022-10-13 riatelab/osrm#114 @JosiahParry
Define "isometric" and "isodistance"
Related to #106 and #113, define isochrone and isodistance.

When reviewing `osrmIsometric()` the title is `"Get Polygons of Isodistances"`. In order to understand what `osrmIsometric()` does, one must know what an "isodistance" is.

A brief 1 - 2 sentences on the definition of each in `#' @details` would be exceptionally informative for a user who is not a domain expert but wants to use the routing and calculation functionality.
