### 2022-10-13 #4574 @JosiahParry
Outstanding documentation improvements:

- https://github.com/riatelab/osrm/issues/106
- https://github.com/riatelab/osrm/issues/114
- https://github.com/riatelab/osrm/issues/113

Error handling improvements (though, I don't think this is necessarily a show stopper):

- https://github.com/riatelab/osrm/issues/110
- https://github.com/riatelab/osrm/issues/112

Once th documentation improvements have been made, I'm happy to tick the last box. I do think that better error handling should be incorporated as well, but I don't necessarily think that takes away from the functionality in a significant way. It just improves the en user experience.
