### 2022-10-13 6a63afd
fix: improve error handling for vector of coordinates out of bound
fix #110

[ci skip]
