### 2022-10-14 #4574 @wcjochem
> Thank you for the feedback @wcjochem.
> I've just tested the installation on 2 Windows machines without problem. I would gladly add more descriptive installation instructions, but I would firstly need a more specific report of the problems (can you tell me more about these problems @wcjochem ?).

@rCarto, thanks for your reply. I have managed to install the package on Windows now. I suspect it was something else out of date preventing the install. I can't reproduce my errors. I've completed the checklist now.
