### 2022-10-18 0912d27
fix: handle empty isochrones
fix #112
