+++ NAMESPACE
@@ -17,5 +17,6 @@ importFrom(sf,st_geometry)
 importFrom(sf,st_intersects)
 importFrom(sf,st_make_grid)
 importFrom(sf,st_sf)
+importFrom(sf,st_sfc)
 importFrom(sf,st_transform)
 importFrom(sf,st_union)
