### 2022-10-18 9a14ffb
fix: add isochrone and isodistance definitions in paper, README, and function documentation; rename osrmIsometric() to osrmIsodistance()
fix #113 fix #106 fix #114
