### 2022-10-18 #4574 @rCarto
Hi @JosiahParry ,
I've added a short definition of isochrones in README, paper and function doc. I understand that these definitions could be useful to non-specialists but I won't go further, isochrone maps are quite standard objects in analyses using travel times.
I've also addressed the error handling issues you raised.
