### 2022-10-18 #4574 @elbeejay
Big thanks to our reviewers @JosiahParry, @mikemahoney218, @wcjochem for providing thoughtful and constructive review comments to this submission. @rCarto has been able to make revisions in response to those reviews, and I think we've achieved some consensus on the final state of the code and manuscript. So again, thank you to our reviewers; this process obviously doesn't exist without you.

@rCarto there are a few more steps we will go through before we can get `osrm` published. First, I will be having the bot check the references in the paper, then I will go through the paper and software once myself to see if there are any typos and things that slipped through the review process. Once that is done I will ping you and ask you to complete a few small tasks (archiving and getting a DOI) so that we can proceed to publication.
