### 2022-10-18 #4574 @elbeejay
@rCarto after looking through the package and paper once more, I have a few suggestions:

- Delete stale branches in repository "mapiso" and "nearest"
- Line 22: I suggest replacing the word "fitted" with "suited"
- Line 25: I believe "interfaces" should be the singular "interface" here
- Line 48: I would add a clause, something like "uses OSRM, however it exposes"
- Line 50: I believe it should be "to give access" and then "and handle spatial" in the next line
- Line 54: Consider "shipped with the package, and contains a sample of" rather than "it is" as written
- Line 79: You can drop "through it" and have the sentence read "The route is 12.3 kilometers long and takes 21.1 minutes to drive."
- I apologize for not noting this earlier, but it would be good to have captions for each figure that provide a brief description of its contents. I realize they are described in the text and the code to produce them is immediately above each figure, but I think it may help readability and allow someone to quickly peruse the figures of the paper and know what they are looking it.

Let me know if you have any questions, feel free to implement any of these suggestions (or none of them). Just let me know when you are finished making any final changes and I will let you know what the next steps are.
