### 2022-10-19 0ab3715
docs: address modification suggested  in https://github.com/openjournals/joss-reviews/issues/4574#issuecomment-1282867516
