### 2022-10-19 #4574 @rCarto
@elbeejay , I've modified the paper following your suggestions. All is ok on my side, I'll wait for your instructions.

@elbeejay , @wcjochem , @JosiahParry and @mikemahoney218
Your careful reading of the paper/doc/code and your precise suggestions helped me raise the package to a much higher quality level. So, **thank you very much** for your work on this review!
