### 2022-10-19 #4574 @elbeejay
Great, thanks for making those changes. At this time, @rCarto if you could do the following:

- [ ] Release the revised software on CRAN (last release was quite a while ago, would be nice to have a CRAN release associated with the JOSS version)

Then we will have the metadata and archiving all set, and I'll be able to recommend this for publication in JOSS. Thanks!
