### 2022-10-19 dd00b34
test: switch off tests of demo and local servers
