### 2022-10-20 56e6fa3
docs: update cff and codemeta
[ci skip]
