### 2022-10-20 ffe41fb
test: remove a tinytest:: call from test suite
