### 2022-10-20 ccc58c8
fix: remove unneeded packages in Suggests, update codemeta and cff
