+++ DESCRIPTION
@@ -23,7 +23,7 @@ Imports:
     sf
 Depends:
     R (>= 3.5.0)
-Suggests: mapsf, lwgeom, tinytest, covr, maptiles
+Suggests: mapsf, tinytest, covr
 URL: https://github.com/riatelab/osrm
 BugReports: https://github.com/riatelab/osrm/issues
 Encoding: UTF-8
