### 2022-10-20 #4574 @rCarto
- [x] Tagged release of `osrm` on GitHub: v4.0.0 - https://github.com/riatelab/osrm/releases/tag/v4.0.0
- [x] DOI of the archived version on Zenodo: 10.5281/zenodo.7228376 - [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7228376.svg)](https://doi.org/10.5281/zenodo.7228376)
- [ ] CRAN Release
