### 2022-10-25 #4574 @kthyng
Hi @rCarto! Things are looking ready to go except — please go through your references and fix up the capitalization. For example, "r" is not capitalized several places, but please go through each carefully to be sure they all look perfect.
