### 2022-10-25 #4574 @kthyng
oh and you can preserve capitalization in your .bib file with {} around characters/strings.
