### 2022-10-26 054c1e7
docs: fix citation capitalisation in paper
see https://github.com/openjournals/joss-reviews/issues/4574#issuecomment-1290612950
