+++ paper.bib
@@ -18,25 +18,6 @@
 }
 
 
-@article {dm0088,
-	author = "Huber, S.",
-	author = "Rust, C.",
-	title = "Calculate travel time and distance with OpenStreetMap data using the Open Source Routing Machine (OSRM)",
-	journal = "Stata Journal",
-	publisher = "Stata Press",
-	address = "College Station, TX",
-	volume = "16",
-	number = "2",
-	year = "2016",
-	pages = "416-423(8)",
-	url = "//www.stata-journal.com/article.html?article=dm0088"
-}
-
-
-
-
-
-
 
 @article{https://doi.org/10.1111/cobi.13326,
 author = {Hickisch, R. and Hodgetts, T. and Johnson, P. J. and Sillero-Zubiri, C. and Tockner, K. and Macdonald, D.W.},
@@ -106,7 +87,7 @@ url={https://revistatransportes.org.br/anpet/article/view/2385},
 DOI={10.14295/transportes.v29i2.2385}, 
 number={2}, 
 journal={TRANSPORTES}, 
-author={Fernandes Barroso, Joana Maia and Albuquerque Oliveira, João Lucas and de Oliveira Neto, Francisco Moraes}, 
+author={Fernandes Barroso, Joana Maia and Albuquerque Oliveira, João Lucas and De Oliveira Neto, Francisco Moraes}, 
 year={2021}, 
 month={ago.}, 
 pages={2385} 
@@ -125,7 +106,7 @@ language = {ger},
 number = {4},
 pages = {63-78},
 publisher = {Institut der deutschen Wirtschaft (IW)},
-title = {Ein hedonischer Mietpreisindex f\"{u}r Einzelhandelsimmobilien in Deutschland: Methodik und erste Ergebnisse; A hedonic rental price index for retail properties in Germany: Methodology and first results},
+title = {Ein hedonischer Mietpreisindex f\"{u}r Einzelhandelsimmobilien in Deutschland: Methodik und erste Ergebnisse},
 url = {http://hdl.handle.net/10419/251180},
 volume = {48},
 year = {2021}
@@ -186,7 +167,7 @@ doi = {10.1108/IJHMA-02-2018-0017}
 
 @Article{ijerph18073813,
 AUTHOR = {Rice, Whitney S. and Labgold, Katie and Peterson, Quita Tinsley and Higdon, Megan and Njoku, Oriaku},
-TITLE = {Sociodemographic and Service Use Characteristics of Abortion Fund Cases from Six States in the U.S. Southeast},
+TITLE = {Sociodemographic and Service Use Characteristics of Abortion Fund Cases from Six States in the {U.S.} Southeast},
 JOURNAL = {International Journal of Environmental Research and Public Health},
 VOLUME = {18},
 YEAR = {2021},
@@ -264,7 +245,7 @@ author = {Evan M. Chen and Aneesha Ahluwalia and Ravi Parikh and Kristen Nwanyan
 
 @article{doi:10.1177/23998083211040519,
 author = {Patrick Ballantyne and Alex Singleton and Les Dolega and Kevin Credit},
-title ={A framework for delineating the scale, extent and characteristics of American retail centre agglomerations},
+title ={A framework for delineating the scale, extent and characteristics of {A}merican retail centre agglomerations},
 journal = {Environment and Planning B: Urban Analytics and City Science},
 volume = {49},
 number = {3},
@@ -278,7 +259,7 @@ eprint = {https://doi.org/10.1177/23998083211040519}
 
 
 @article{WALKER2021127097,
-title = {Assessing challenges and opportunities for schools’ access to nature in England},
+title = {Assessing challenges and opportunities for schools’ access to nature in {E}ngland},
 journal = {Urban Forestry & Urban Greening},
[…]
